requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
faiss-cpu
numpy